LayoutTests/imported/w3c/web-platform-tests/pointerevents/compat/pointerevent_touch_target_after_pointerdown_target_removed.tentative-expected.txt


PASS After a pointerdown listener removes its target
PASS After a pointerdown listener removes but appends the target to same position again
PASS After a pointerdown listener moves the target to different position
PASS After a touchstart listener removes its target
PASS After a touchstart listener removes but appends the target to same position again
PASS After a touchstart listener moves the target to different position
FAIL After a pointerdown listener removes its target, touch events should be fired on the touchstart target even though an orphan and pointer events should be fired on the parent assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: touchstart, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointermove, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: pointerup, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchend, target: Element node <div id=\"pointerDownTarget\"></div> } ]" but got "[]"
FAIL After a pointerdown listener removes its target, click event should be fired on the pointerdown target parent assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: touchstart, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointermove, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: pointerup, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchend, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: click, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... } ]" but got "[]"
FAIL After a pointerdown listener removes its target, touchmove event should be fired on the pointerdown target assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: touchstart, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointermove, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchmove, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointerup, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchend, target: Element node <div id=\"pointerDownTarget\"></div> } ]" but got "[]"
PASS After a pointerdown listener removes but appends the target to same position again, touch events and pointer events should be fired on the pointerdown target
PASS After a pointerdown listener removes but appends the target to same position again, click event should be fired on the pointerdown target
PASS After a pointerdown listener removes but appends the target to same position again, touchmove event should be fired on the pointerdown target
FAIL After a pointerdown listener moves the target to different position, touch events should be fired on the pointerdown target, but pointer events should be fired on the pointerdown target parent assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: touchstart, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointermove, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: pointerup, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchend, target: Element node <div id=\"pointerDownTarget\"></div> } ]" but got "[]"
FAIL After a pointerdown listener moves the target to different position, click event should be fired on the pointerdown target parent assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: touchstart, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointermove, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: pointerup, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchend, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: click, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... } ]" but got "[]"
FAIL After a pointerdown listener moves the target to different position, touchmove event should be fired on the pointerdown target parent assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: touchstart, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointermove, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchmove, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointerup, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchend, target: Element node <div id=\"pointerDownTarget\"></div> } ]" but got "[]"
FAIL After a touchstart listener removes its target, touch events should be fired on the touchstart target even though an orphan and pointer events should be fired on the parent assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: touchstart, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointermove, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: pointerup, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchend, target: Element node <div id=\"pointerDownTarget\"></div> } ]" but got "[]"
FAIL After a touchstart listener removes its target, click event should be fired on the pointerdown target parent assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: touchstart, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointermove, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: pointerup, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchend, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: click, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... } ]" but got "[]"
FAIL After a touchstart listener removes its target, touchmove event should be fired on the pointerdown target assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: touchstart, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointermove, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchmove, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointerup, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchend, target: Element node <div id=\"pointerDownTarget\"></div> } ]" but got "[]"
PASS After a touchstart listener removes but appends the target to same position again, touch events and pointer events should be fired on the pointerdown target
PASS After a touchstart listener removes but appends the target to same position again, click event should be fired on the pointerdown target
PASS After a touchstart listener removes but appends the target to same position again, touchmove event should be fired on the pointerdown target
FAIL After a touchstart listener moves the target to different position, touch events should be fired on the pointerdown target, but pointer events should be fired on the pointerdown target parent assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: touchstart, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointermove, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: pointerup, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchend, target: Element node <div id=\"pointerDownTarget\"></div> } ]" but got "[]"
FAIL After a touchstart listener moves the target to different position, click event should be fired on the pointerdown target parent assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: touchstart, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointermove, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: pointerup, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchend, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: click, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... } ]" but got "[]"
FAIL After a touchstart listener moves the target to different position, touchmove event should be fired on the pointerdown target parent assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: touchstart, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointermove, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchmove, target: Element node <div id=\"pointerDownTarget\"></div> }, { type: pointerup, target: Element node <div id=\"pointerDownTargetParent\"><div id=\"pointerDownTar... }, { type: touchend, target: Element node <div id=\"pointerDownTarget\"></div> } ]" but got "[]"
